feat: Add resources (steps) info in workload workload get
Adding resource info in workload get command. There is a list which shows the name of each resource, its status and its age.
The status is taken from the condition of the resource which type is 'Ready'.

According to Cartographer:
'Conditions describing this resource's reconcile
state. The top level condition is of type `Ready`, and follows
these Kubernetes conventions:
https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#typical-status-properties'

There are going to be some follow up commits changing the current output of workload get command.
One to display supply chain info and other to display issues in workload

Signed-off-by: Wendy Arango <warango@vmware.com>