feat: Show supply chain info in workload get
In this feat, the way the workload status used to be displayed changes to show specifically the supply chain info (name) and the workload status on its Ready condition.
If there is no supply chain reference, but there is a Ready condition that surfaces the workload status, on row "name", `<none>` is going to be displayed.
If there is no Ready condition, but a supply chain reference, then rows "ready" and "last update" are left blank.
If there are no conditions and no supply chain reference, a message indicating this situation is displayed.

Follow-up of PR #99 for issue #61

Signed-off-by: Wendy Arango <warango@vmware.com>