fix: Add max log requests for stern tailer
With Stern new version (v1.24.0), a new flag was introduced to
decide which is the maximum number of concurrent requests.
In Apps Plugin code, it never gets to have a default number other than 0,
that's why, if not specified in Stern configuration, logs would never appear
since there were not available requests to do.
A number of the maximum Int16 (32767) was set as the max log requests
for Stern, which means there can be up to that amount of pods logging at the
same time for Apps Plugin `workload tail` or
`workload create/update/apply --tail` to work.

Another change introduced in Stern was the `ALL_STATES` option for containers
state filter. It was also added in Apps Plugin so all containers are shown
(`running,waiting,terminated`).

Signed-off-by: Wendy Arango <warango@vmware.com>